Created main() class and added shebang
So two quick things - the shebang (first line that starts with #!) allows linux to recognise what interpreter to use - windows already knows this through file associations but if you want this to be portable between operating systems this is good practise and it costs you nothing.  99% sure it works on windows too but test it first - if not just delete it.

Secondly and perhaps more importantly the ``def main()
  ...
  ...

if __name__ == '__main__':
  main()
```
Block is useful because now that class you have created can be imported in to another script and the lines of the script in the main block will not be run.  Now you can use that class in any other file just by importing it and you will never have to rewrite it.

References:
Part 1: https://stackoverflow.com/questions/7574453/shebang-notation-python-scripts-on-windows-and-linux
Part 2: https://stackoverflow.com/questions/419163/what-does-if-name-main-do

Hit me up if anything doesn't make sense.